RDKB-43069 : Fix libwebconfig build errors on XE2
Reason for change: libwebconfig builds were failing on XE2 platform
                   due to warnings being treated as errors
Test Procedure: Ensure no build failures on pods and XBs

Risks: Low
(cherry picked from commit 1a8faa8852bd0aac0026ba0039de96606c1153f1)

+++ source/utils/wifi_util.c
@@ -717,7 +717,7 @@ char *get_formatted_time(char *time)
 
     strftime(tmp, 128, "%y%m%d-%T", tm_info);
 
-    snprintf(time, 128, "%s.%06ld", tmp, tv_now.tv_usec);
+    snprintf(time, 128, "%s.%06lld", tmp, (long long)tv_now.tv_usec);
     return time;
 }
